fix: users could delete and update any movies

+++ src/movies/movies.controller.ts
@@ -7,15 +7,19 @@ import {
   Param,
   Post,
   Put,
+  Req,
   Res,
   NotFoundException,
+  ForbiddenException,
 } from '@nestjs/common';
 import { MovieIdDto } from './movies.dto.movie-id';
 import { CreateMovieDto } from './movies.dto.create-movie';
+import { CreateMovieRawDto } from './movies.dto.create-movie-raw';
 import { UpdateMovieDto } from './movies.dto.update-movie';
 import { MovieService } from './movies.service';
 import { ResMovie } from './movies.response.class';
 import { Response } from 'express';
+import { RequestWithUser } from 'src/auth/auth.request-with-user';
 import { Public } from 'src/auth/auth.decorator.public-route';
 
 @Controller('movies')
@@ -24,9 +28,14 @@ export class MovieController {
 
   @Post()
   async createMovie(
+    @Req() request: RequestWithUser,
     @Res() response: Response,
-    @Body() createMovieDto: CreateMovieDto,
+    @Body() createMovieRawDto: CreateMovieRawDto,
   ) {
+    const createMovieDto = new CreateMovieDto(
+      createMovieRawDto,
+      request.user.sub,
+    );
     const newMovie = await this.movieService.createMovie(createMovieDto);
     return response.status(HttpStatus.CREATED).json({
       message: 'Movie has been created successfully',
@@ -37,22 +46,27 @@ export class MovieController {
 
   @Put('/:id')
   async updateMovie(
+    @Req() request: RequestWithUser,
     @Res() response: Response,
     @Param() movieIdDto: MovieIdDto,
     @Body() updateMovieDto: UpdateMovieDto,
   ) {
-    const existingMovie = await this.movieService.updateMovie(
-      movieIdDto,
-      updateMovieDto,
-    );
+    const existingMovie = await this.movieService.getMovie(movieIdDto);
     if (!existingMovie) {
       throw new NotFoundException(
         MovieController.NotFoundMessage(movieIdDto.id),
       );
     }
+    if (existingMovie.creatorUserId.toString() !== request.user.sub) {
+      throw new ForbiddenException();
+    }
+    const updatedMovie = await this.movieService.updateMovie(
+      movieIdDto,
+      updateMovieDto,
+    );
     return response.status(HttpStatus.OK).json({
       message: 'Movie has been successfully updated',
-      result: new ResMovie(existingMovie),
+      result: new ResMovie(updatedMovie),
       statusCode: HttpStatus.OK,
     });
   }
@@ -86,15 +100,20 @@ export class MovieController {
 
   @Delete('/:id')
   async deleteMovie(
+    @Req() request: RequestWithUser,
     @Res() response: Response,
     @Param() movieIdDto: MovieIdDto,
   ) {
-    const deletedMovie = await this.movieService.deleteMovie(movieIdDto);
-    if (!deletedMovie) {
+    const existingMovie = await this.movieService.getMovie(movieIdDto);
+    if (!existingMovie) {
       throw new NotFoundException(
         MovieController.NotFoundMessage(movieIdDto.id),
       );
     }
+    if (existingMovie.creatorUserId.toString() !== request.user.sub) {
+      throw new ForbiddenException();
+    }
+    const deletedMovie = await this.movieService.deleteMovie(movieIdDto);
     return response.status(HttpStatus.OK).json({
       message: 'Movie deleted successfully',
       result: new ResMovie(deletedMovie),
